split addUser
