+++ Snippets/websockets_02.swift
@@ -7,10 +7,8 @@ import HummingbirdWSCompression
 import Logging
 import ServiceLifecycle
 
-// snippet.ConnectionManager
 struct ConnectionManager: Service {
 
-    // ...
     // snippet.hide
     typealias OutputStream = AsyncChannel<WebSocketOutboundWriter.OutboundFrame>
 
@@ -44,7 +42,6 @@ struct ConnectionManager: Service {
             await send("\(name) left")
         }
     }
-    // snippet.show
 
     let connectionStream: AsyncStream<Connection>
     let connectionContinuation: AsyncStream<Connection>.Continuation
@@ -57,54 +54,11 @@ struct ConnectionManager: Service {
         self.logger = logger
     }
 
-    func run() async {
-        await withGracefulShutdownHandler {
-            await withDiscardingTaskGroup { group in
-                let outboundCounnections = OutboundConnections()
-                // 1.
-                for await connection in connectionStream {
-                    group.addTask {
-                        logger.info("add connection", metadata: [
-                            "name": .string(connection.name)
-                        ])
-                        // 2.
-                        await outboundCounnections.add(
-                            name: connection.name,
-                            outbound: connection.outbound
-                        )
+    func run() async {}
 
-                        do {
-                            // 3.
-                            for try await input in connection.inbound.messages(
-                                maxSize: 1_000_000
-                            ) {
-                                guard case .text(let text) = input else {
-                                    continue
-                                }
-                                let output = "[\(connection.name)]: \(text)"
-                                logger.debug("Output", metadata: [
-                                    "message": .string(output)
-                                ])
-                                // 4.
-                                await outboundCounnections.send(output)
-                            }
-                        } catch {}
+    // snippet.show
 
-                        logger.info("remove connection", metadata: [
-                            "name": .string(connection.name)
-                        ])
-                        // 5.
-                        await outboundCounnections.remove(name: connection.name)
-                        connection.outbound.finish()
-                    }
-                }
-                group.cancelAll()
-            }
-        } onGracefulShutdown: {
-            connectionContinuation.finish()
-        }
-    }
-
+    // ...
     func addUser(
         name: String,
         inbound: WebSocketInboundStream,
@@ -120,84 +74,3 @@ struct ConnectionManager: Service {
         return outputStream
     }
 }
-// snippet.end
-
-// snippet.buildApplication
-func buildApplication(
-    _ arguments: some AppArguments
-) async throws -> some ApplicationProtocol {
-    var logger = Logger(label: "WebSocketChat")
-    logger.logLevel = .trace
-
-    // 1.
-    let router = Router()
-    router.middlewares.add(LogRequestsMiddleware(.debug))
-    router.middlewares.add(FileMiddleware(logger: logger))
-
-    // 2.
-    let connectionManager = ConnectionManager(logger: logger)
-    // 3.
-    let wsRouter = Router(context: BasicWebSocketRequestContext.self)
-    wsRouter.middlewares.add(LogRequestsMiddleware(.debug))
-    // 4.
-    wsRouter.ws("chat") { request, _ in
-        guard request.uri.queryParameters["username"] != nil else {
-            return .dontUpgrade
-        }
-        return .upgrade([:])
-    // 5.
-    } onUpgrade: { inbound, outbound, context in
-        guard let name = context.request.uri.queryParameters["username"] else {
-            return
-        }
-        let outputStream = connectionManager.addUser(
-            name: String(name),
-            inbound: inbound,
-            outbound: outbound
-        )
-        for try await output in outputStream {
-            try await outbound.write(output)
-        }
-    }
-
-    // 6.
-    var app = Application(
-        router: router,
-        server: .http1WebSocketUpgrade(
-            webSocketRouter: wsRouter,
-            configuration: .init(extensions: [.perMessageDeflate()])
-        ),
-        configuration: .init(
-            address: .hostname(arguments.hostname, port: arguments.port)
-        ),
-        logger: logger
-    )
-    // 7.
-    app.addServices(connectionManager)
-    return app
-}
-// snippet.end
-
-// snippet.entrypoint
-// 1.
-protocol AppArguments {
-    var hostname: String { get }
-    var port: Int { get }
-}
-
-// 2.
-@main
-struct HummingbirdArguments: AppArguments, AsyncParsableCommand {
-    @Option(name: .shortAndLong)
-    var hostname: String = "127.0.0.1"
-
-    @Option(name: .shortAndLong)
-    var port: Int = 8080
-
-    func run() async throws {
-        // 3.
-        let app = try await buildApplication(self)
-        try await app.runService()
-    }
-}
-// snippet.end
